requirements.txt
pillow
python=3.6